fix(mysql): faz o ddl-auto=validate passar contra um MySQL real
Item E5 do backlog confirmado: o perfil mysql nunca tinha rodado contra um
MySQL de verdade, e nao subia. O validate compara a FAMILIA do tipo JDBC de
cada coluna com a do atributo, e 54 das 133 colunas mapeadas divergiam.

No Oracle nada disso aparece — o dialeto mapeia CHAR e VARCHAR ambos para
VARCHAR2, nao existe ENUM nativo, e NUMBER absorve qualquer inteiro. Como a
suite roda em H2 MODE=Oracle, nenhum dos tres casos era exercitado.

As tres classes, e por que cada correcao e do jeito que e:

  33 colunas de UUID — o comum.properties pede preferred_uuid_jdbc_type=CHAR
  para o Hibernate nao gravar UUID como BINARY(16), mas CHAR faz ele esperar
  char(36) e as migrations criam VARCHAR(36). Trocado para VARCHAR no perfil
  mysql: continua sendo texto, resolve o BINARY(16) e bate com o DDL. Fica no
  perfil, e nao no comum.properties, porque no Oracle o padrao ja esta certo.

  14 colunas de enum — desde o Hibernate 6.2, @Enumerated(STRING) com
  MySQLDialect mapeia para o ENUM nativo do MySQL. Corrigido pelo lado da
  configuracao, e nao do banco: com ENUM nativo, acrescentar um valor ao enum
  Java passaria a exigir ALTER TABLE.

  7 colunas booleanas — estas nao tem correcao por propriedade. O
  NumericBooleanConverter faz o atributo chegar ao JDBC como Integer, entao o
  Hibernate espera INTEGER mesmo o campo Java sendo boolean. TINYINT vira INT
  em usuario.ativo (V3), servico.ativo, alerta_clinico.ativo, animal.castrado
  (V6), acesso_historico.nivel, acesso_historico.emergencial e
  animal.resumo_seguranca_ativo (V7), preservando nulidade e default de cada
  uma. So no conjunto mysql: no oracle continua NUMBER(1), que valida.

Editado no lugar em vez de uma V8 com ALTER: nada esta publicado, entao o
checksum novo nao afeta ninguem, e o estado final fica legivel. Quem tiver
banco local precisa recriar ou rodar flyway repair.

Verificado com banco criado do zero: Flyway aplica V1..V7, a aplicacao sobe
com validate ligado em 9.7s, e o fluxo POST /animais -> POST /lembretes da API
.NET sobre o mesmo schema responde 201 com castrado fazendo round-trip pela
coluna INT. MigrationsMySqlTest continua em 16/16.

+++ src/main/resources/db/migration/mysql/V3__usuario_e_perfis.sql
@@ -18,11 +18,13 @@
 -- tres sao as unicas do projeto inteiro que nao sao renomeacao
 -- mecanica de tipo, entao vale saber por que cada uma e assim:
 --
---   ativo NUMBER(1) -> TINYINT
+--   ativo NUMBER(1) -> INT
 --       Usuario.ativo e boolean anotado com NumericBooleanConverter,
 --       ou seja, o Hibernate grava 0/1 num inteiro -- nao um BOOLEAN
---       nativo. TINYINT e o inteiro de 1 byte do MySQL e o BOOLEAN do
---       MySQL e apelido dele, entao os dois lados falam a mesma coisa.
+--       nativo. Precisa ser INT, e nao TINYINT: o conversor faz o
+--       atributo chegar ao JDBC como Integer, e o ddl-auto=validate
+--       compara familia de tipo JDBC. TINYINT e INTEGER sao familias
+--       diferentes, e a divergencia impede o boot da aplicacao.
 --
 --   tentativas_falhas NUMBER(3) -> INT
 --       O campo e int em Java. SMALLINT caberia de sobra (o limite de
@@ -45,7 +47,7 @@ CREATE TABLE usuario (
     email             VARCHAR(200) NOT NULL,
     senha             VARCHAR(100) NOT NULL,
     perfil            VARCHAR(20)  NOT NULL,
-    ativo             TINYINT      DEFAULT 1 NOT NULL,
+    ativo             INT          DEFAULT 1 NOT NULL,
     tentativas_falhas INT          DEFAULT 0 NOT NULL,
     bloqueado_ate     DATETIME,
     tutor_id          VARCHAR(36),
